backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
sqlalchemy==2.0.21
pydantic[email]==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.6
email-validator==2.1.0
slowapi==0.1.9
alembic==1.13.1

# Testing
pytest==8.1.1
httpx==0.27.0
